backend/requirements.txt
fastapi[standard]
mysql-connector-python
sqlalchemy
sqlmodel
pydantic